hand-gesture-recognition/requirements.txt
tensorflow>=2.20.0
opencv-python>=4.5.0
numpy>=1.26.0
matplotlib>=3.5.0
scikit-learn>=1.7.0
seaborn>=0.12.0
Pillow>=9.0.0

# Optional: For edge devices with limited resources
# Uncomment if using TFLite Runtime instead of full TensorFlow
# tflite-runtime>=2.14.0

